refactor findAllFilePaths
