+++ _internal/pathResolve.js
@@ -1,0 +1,6 @@
+const path = require('path')
+
+// pathResolve(...args) -> resolvedPath string
+const pathResolve = (...args) => path.resolve(...args)
+
+module.exports = pathResolve
