+++ dist-test.js
@@ -1,24 +1,4 @@
-const fs = require('fs')
-const path = require('path')
-
-const pathResolve = (...args) => path.resolve(...args)
-
-const findAllFilePaths = async function (path) {
-  const files = await fs.promises.readdir(path, {
-    withFileTypes: true,
-    recursive: true,
-  })
-  const result = []
-  for (const file of files) {
-    if (file.isDirectory()) {
-      const filePaths = await findAllFilePaths(pathResolve(path, file.name))
-      result.push(...filePaths)
-    } else {
-      result.push(pathResolve(path, file.name))
-    }
-  }
-  return result
-}
+const findAllFilePaths = require('./_internal/findAllFilePaths')
 
 const distTest = async function main() {
   const filePaths = await findAllFilePaths(`${__dirname}/dist`)
